test: extract with_insertion_fill rejection test to unit tier
Move test_with_insertion_fill_rejects_when_no_tracks_active to
tests/unit/dataset/test_with_insertion_fill.py. The test uses
gvl.get_dummy_dataset() (gvl-shipped helper, not a project toy-
dataset fixture) and only checks API-level rejection behavior.
The 2 remaining tests in the integration file exercise the full
reconstruction call path and stay in integration.

Phase 5 dataset polymorphism component — minimal scope (test_ds_indexing
deferred pending a make_dataset consumer; see status doc).
